requirements.txt
# FastAPI and Core Dependencies
fastapi
uvicorn
pydantic
pydantic-settings

# Database
sqlalchemy
pymysql
cryptography
alembic

# Document Processing
python-docx
Pillow
pypdf
reportlab

# Email
Jinja2
aiosmtplib

# Payment Processing
stripe

# AI / LLM
openai

# Utilities
python-multipart
python-jose
passlib
bcrypt
pyyaml
